requirements.txt
bandit
coverage
deepdiff
django-cors-headers
django-cron
django-crontab
django-environ
django-filter
django-phonenumber-field
django-wkhtmltopdf
djangorestframework-simplejwt
drf-writable-nested
drf-yasg
gunicorn
install
phonenumberslite
pip-autoremove
pip-chill
psycopg2
psycopg2-binary
pylint
pylint-venv
pyopenssl
pytest-django
pytest-xdist
python-dateutil
python-dotenv
service-identity
toml
xhtml2pdf
